TTTable: fix for a stupid gotcha trying to make the '>' appear in a table cell
This functionality is implemented in TTTableLinkedItemCell.m, in setObject. Remember, TTTableLinkedItemCell is a subclass of UITableViewCell, and so this is where you actually control what kind of table cell is created. Also remember that TTTableLinkedItemCell is the class that is created when you add a TTTableLinkedItem (or one of it's numerous subclasses) to the table in your TTTableDatasource class.

The rules here are:
- if the TTTableLinkedItem has an accessoryURL, then:
a) accessoryType => UITableViewCellAccessoryDetailDisclosureButton
- if the navigationMode for the URL associated with the TTTableLinkedItem is (TTNavigationModeCreate || TTNavigationModeShare), then:
b)accessoryType => UITableViewCellAccessoryDisclosureIndicator

This is sensible, but what if you set a URL for the TTTableLinkedItem, and it still doesn't appear? Probably it is returning TTNavigationModeNone, which is in turn b/c TTNavigator is not usefully documented.

Here's one way to cause that: use TTURLMap from:toObject:selector:, like this:

TTNavigator* navigator = [TTNavigator navigator];
TTURLMap* map = navigator.URLMap;
[map from:@"drync://tag_custom_location"
	   toObject:self
	   selector:@selector(tag_custom_location)
	   ];

Note: if you use other variants in TTURLMap, like the toViewController/toSharedViewController/toModalViewController variants, then the issue doesn't arise. The problem is exactly if you want to carefully control the creation of the view controller, perhaps because you need to interface to an old-style, non-three20 view.

Ended out patching TTUrlMap as follows to allow TTNavigationMode to be specified, if you care to.